Refactor: Move GitHub-related classes to repositories package
